Rename DemoControls components to DocsPlaygroundControls
